requirements.txt
asgiref==3.3.1
autopep8==1.5.6
boto3==1.17.33
botocore==1.20.33
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Django==3.1.4
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
newsapi-python==0.2.6
numpy==1.19.0
opencv-python==4.3.0.36
optional-django==0.3.0
Pillow==8.1.2
praw==7.2.0
prawcore==2.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
react==4.3.0
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
style==1.1.0
toml==0.10.2
tzlocal==2.1
update==0.0.1
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.2
websocket-client==0.58.0
Werkzeug==1.0.1
whitenoise==5.2.0
